backend/requirements.txt
arize
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
python-dotenv
requests
openai
pandas
